Workaround buildkite-agent absolute path issue (#13)
Works around buildkite-agent absolute path problem - doesn't understand drive-rooted paths on windows. See also buildkite/agent#881.

Makes the start parameters configurable. Also delete the configuration file from the zip file, because the agent prefers to find its configuration next-door to the binary ahead of any of the search paths, and if the stub is in place it'll use that in preference to our rendered version.